CL_ActionEvent in SP had an unsigned long type in argument which is 64 bit on 64-bit non-windows. Its okay to be 32 bit.

+++ code/client/cl_keys.cpp
@@ -1283,7 +1283,7 @@ void CL_InitKeyCommands( void ) {
 }
 
 
-void CL_ActionEvent(int key, qboolean down, ulong time)
+void CL_ActionEvent(int key, qboolean down, unsigned time)
 {
 	const char	*kb;
 	char		cmd[1024];
